# opensovd-core/sovd-server/openapi.yaml
openapi: 3.1.0
info:
  title: sovd-server
  description: HTTP/REST SOVD server for the Eclipse OpenSOVD core stack
  license:
    name: Apache-2.0
    identifier: Apache-2.0
  version: 0.1.0
paths:
...
  /sovd/v1/components/{component_id}/bulk-data:
    post:
      tags:
      - bulk-data
      summary: '`POST /sovd/v1/components/{component_id}/bulk-data` â€” start a transfer.'
      operationId: startBulkDataTransfer
      parameters:
      - name: component_id
        in: path
        description: Stable component identifier
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BulkDataTransferRequest'
        required: true
      responses:
        '201':
          description: Transfer created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkDataTransferCreated'
        '400':
          description: Invalid manifest
        '404':
          description: Component not found
        '422':
          description: Component does not support bulk-data
  /sovd/v1/components/{component_id}/bulk-data/{transfer_id}:
    put:
      tags:
      - bulk-data
      summary: |-
        `PUT /sovd/v1/components/{component_id}/bulk-data/{transfer_id}` â€” upload
        one binary chunk.
      operationId: uploadBulkDataChunk
      parameters:
      - name: component_id
        in: path
        description: Stable component identifier
        required: true
        schema:
          type: string
      - name: transfer_id
        in: path
        description: Bulk-data transfer identifier
        required: true
        schema:
          type: string
      requestBody:
        description: Binary chunk payload addressed by Content-Range
        content:
          application/octet-stream:
            schema:
              type: string
        required: true
      responses:
        '204':
          description: Chunk accepted
        '400':
          description: Invalid Content-Range or payload
        '404':
          description: Transfer not found
        '409':
          description: Chunk order conflicts with current transfer state
    delete:
      tags:
      - bulk-data
      summary: '`DELETE /sovd/v1/components/{component_id}/bulk-data/{transfer_id}` â€” abort.'
      operationId: cancelBulkDataTransfer
      parameters:
      - name: component_id
        in: path
        description: Stable component identifier
        required: true
        schema:
          type: string
      - name: transfer_id
        in: path
        description: Bulk-data transfer identifier
        required: true
        schema:
          type: string
      responses:
        '204':
          description: Transfer cancelled
        '404':
          description: Transfer not found
  /sovd/v1/components/{component_id}/bulk-data/{transfer_id}/status:
    get:
      tags:
      - bulk-data
      summary: '`GET /sovd/v1/components/{component_id}/bulk-data/{transfer_id}/status`.'
      operationId: getBulkDataTransferStatus
      parameters:
      - name: component_id
        in: path
        description: Stable component identifier
        required: true
        schema:
          type: string
      - name: transfer_id
        in: path
        description: Bulk-data transfer identifier
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Current transfer status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkDataTransferStatus'
        '404':
          description: Transfer not found
  /sovd/v1/components/{component_id}/data:
    get:
      tags:
      - data-access
      summary: |-
        `GET /sovd/v1/components/{component_id}/data` — list the
        data-metadata catalog.
      description: |-
        # Errors

        Returns 404 if the component is unknown; other
        [`SovdError`](sovd_interfaces::SovdError) values are mapped via
        [`ApiError`].
      operationId: listData
      parameters:
      - name: component_id
        in: path
        description: Stable component identifier
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Data-metadata catalog
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Datas'
        '404':
          description: Component not found
  /sovd/v1/components/{component_id}/data/{data_id}:
    get:
      tags:
      - data-access
      summary: |-
        `GET /sovd/v1/components/{component_id}/data/{data_id}` — read one
        live data value.
      description: |-
        # Errors

        Returns 404 if the component or data id is unknown; other
        [`SovdError`](sovd_interfaces::SovdError) values are mapped via
        [`ApiError`].
      operationId: readData
      parameters:
      - name: component_id
        in: path
        description: Stable component identifier
        required: true
        schema:
          type: string
      - name: data_id
        in: path
        description: Stable data identifier
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Live data value
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SovdReadValue'
        '404':
          description: Component or data id not found
...
components:
  schemas:
...
    BulkDataFailureReason:
      type: string
      description: Enumerated failure reason surfaced by OTA status polling.
      enum:
      - InvalidManifest
      - SignatureInvalid
      - ChunkOutOfOrder
      - FlashWriteFailed
      - PowerLoss
      - AbortRequested
      - Other
    BulkDataState:
      type: string
      description: Bulk-data transfer lifecycle.
      enum:
      - Idle
      - Downloading
      - Verifying
      - Committed
      - Failed
      - Rolledback
    BulkDataTransferCreated:
      type: object
      description: '`POST /components/{id}/bulk-data` success body.'
      required:
      - transfer-id
      - state
      properties:
        state:
          $ref: '#/components/schemas/BulkDataState'
          description: Initial lifecycle state returned to the caller.
        transfer-id:
          type: string
          description: Stable identifier for subsequent `PUT` / `GET status` / `DELETE`.
    BulkDataTransferRequest:
      type: object
      description: '`POST /components/{id}/bulk-data` body.'
      required:
      - manifest
      - image-size
      properties:
        image-size:
          type: integer
          format: int64
          description: Total image size in bytes.
          minimum: 0
        manifest:
          description: |-
            Free-form OTA manifest. Phase 6 requires this to be present even when a
            backend uses only a subset of the fields.
        target-slot:
          type:
          - string
          - 'null'
          description: Optional target-slot preference from the caller.
    BulkDataTransferStatus:
      type: object
      description: '`GET /components/{id}/bulk-data/{transfer-id}/status` body.'
      required:
      - transfer-id
      - state
      - bytes-received
      - total-bytes
      properties:
        bytes-received:
          type: integer
          format: int64
          description: Monotonic byte counter of accepted payload data.
          minimum: 0
        reason:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/BulkDataFailureReason'
            description: Failure reason when `state == Failed`.
        state:
          $ref: '#/components/schemas/BulkDataState'
          description: Current lifecycle state.
        target-slot:
          type:
          - string
          - 'null'
          description: Echo of the optional slot hint when the caller supplied one.
        total-bytes:
          type: integer
          format: int64
          description: Total image size declared when the transfer was created.
          minimum: 0
        transfer-id:
          type: string
          description: Transfer identifier the status belongs to.
...
    DataError:
      type: object
      description: |-
        SOVD partial-error envelope, used when only one element of a structured
        response failed.

        Provenance: `commons/errors.yaml#DataError`.

        `path` is a JSON Pointer pointing at the offending field of the response
        body. `error` describes the failure in more detail.
      required:
      - path
      properties:
        error:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/GenericError'
            description: Detailed error describing why `path` failed.
        path:
          type: string
          description: JSON Pointer to the offending element of the response.
...
    Datas:
      type: object
      description: |-
        Response body for `GET .../components/{component}/data` — the list of
        all data resources an entity provides.

        Provenance: `data/responses.yaml#Datas` (inline response schema).

        The schema is an inline object under the `Datas` response, with two
        fields:

        - `items`: array of `ValueMetadata` (required)
        - `schema`: optional `OpenApiSchema` reference — only populated when
          the client passes `include-schema=true`. The spec defines
          `OpenApiSchema` as an arbitrary `OpenAPI` 3.1 schema subtree pulled in
          via `$ref` from the upstream OAI schema file, so we carry it as a
          free-form `serde_json::Value`. This is **not** a spec-type escape
          hatch — the spec itself keeps this field intentionally open.
      required:
      - items
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ValueMetadata'
          description: |-
            All data resources the entity exposes, each described by its
            `ValueMetadata`.
        schema:
          description: |-
            Optional embedded `OpenAPI` schema describing the response shape.
            Only present when the client requests `include-schema=true`.
...
    GenericError:
      type: object
      description: |-
        SOVD generic error envelope.

        Provenance: `commons/errors.yaml#GenericError`.

        Returned in the body of any non-2xx SOVD response. The `error_code` is
        a SOVD-standardised code; `vendor_code` is filled when the error is
        vendor-specific. `parameters` carries open key/value detail.
      required:
      - error_code
      - message
      properties:
        error_code:
          type: string
          description: SOVD standardised error code.
        message:
          type: string
          description: Human-readable problem description.
        parameters:
          description: Additional key/value detail about the error.
        translation_id:
          type:
          - string
          - 'null'
          description: Identifier for translating `message`.
        vendor_code:
          type:
          - string
          - 'null'
          description: Vendor-specific error code, set when `error_code` is vendor-specific.
...
    SovdReadValue:
      type: object
      description: |-
        Single-value read result returned from `GET .../data/{data-id}`.

        Provenance: `commons/types.yaml#ReadValue`. Rust-named
        `SovdReadValue` for the same reason as [`SovdValue`].
      required:
      - id
      - data
      properties:
        data:
          description: Decoded value payload (`AnyValue` per spec — open).
        errors:
          type:
          - array
          - 'null'
          items:
            $ref: '#/components/schemas/DataError'
          description: Errors encountered while reading parts of `data`.
        id:
          type: string
          description: Stable identifier for the value.
        schema:
          description: Optional embedded JSON Schema describing the response shape.
...
    ValueMetadata:
      type: object
      description: |-
        Metadata describing a data value.

        Provenance: `commons/types.yaml#ValueMetadata`.
      required:
      - id
      - name
      - category
      properties:
        category:
          type: string
          description: |-
            Category, e.g. `currentData`, `identData`, `sysInfo`, or
            `x-<oem>-…`. The spec restricts the string to URL-safe characters
            matching `^[A-Za-z0-9_-]+$`; we do not enforce that at type level.
        groups:
          type:
          - array
          - 'null'
          items:
            type: string
          description: Group identifiers the value belongs to.
        id:
          type: string
          description: Stable identifier for the value (e.g. `"DriverWindow"`).
        name:
          type: string
          description: Human-readable name.
        tags:
          type:
          - array
          - 'null'
          items:
            type: string
          description: Tags attached to this value.
        translation_id:
          type:
          - string
          - 'null'
          description: Identifier for translating `name`.
...